Extract `EditViewOptionalFeaturesMixin` from generic `EditView`
