+++ wagtail/admin/views/generic/models.py
@@ -8,7 +8,6 @@ from django.forms import Form
 from django.http import Http404, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, redirect
 from django.urls import reverse
-from django.utils import timezone
 from django.utils.translation import gettext as _
 from django.utils.translation import gettext_lazy
 from django.views.generic import TemplateView
@@ -22,12 +21,11 @@ from wagtail.actions.unpublish import UnpublishAction
 from wagtail.admin import messages
 from wagtail.admin.forms.search import SearchForm
 from wagtail.admin.panels import get_edit_handler
-from wagtail.admin.templatetags.wagtailadmin_tags import user_display_name
 from wagtail.admin.ui.tables import Column, Table, TitleColumn, UpdatedAtColumn
 from wagtail.admin.utils import get_valid_next_url_from_request
 from wagtail.log_actions import log
 from wagtail.log_actions import registry as log_registry
-from wagtail.models import DraftStateMixin, RevisionMixin
+from wagtail.models import DraftStateMixin
 from wagtail.models.audit_log import ModelLogEntry
 from wagtail.search.backends import get_search_backend
 from wagtail.search.index import class_is_indexed
@@ -493,31 +491,26 @@ class EditView(
     success_message = None
     error_message = None
     submit_button_label = gettext_lazy("Save")
-    actions = ["edit", "publish"]
+    actions = ["edit"]
 
     def setup(self, request, *args, **kwargs):
         super().setup(request, *args, **kwargs)
         self.action = self.get_action(request)
-        self.revision_enabled = self.model and issubclass(self.model, RevisionMixin)
-        self.draftstate_enabled = self.model and issubclass(self.model, DraftStateMixin)
 
     def get_action(self, request):
-        for action in self.actions:
+        for action in self.get_available_actions():
             if request.POST.get(f"action-{action}"):
                 return action
         return "edit"
+
+    def get_available_actions(self):
+        return self.actions
 
     def get_object(self, queryset=None):
         if "pk" not in self.kwargs:
             self.kwargs["pk"] = self.args[0]
         self.kwargs["pk"] = unquote(str(self.kwargs["pk"]))
-        self.live_object = super().get_object(queryset)
-
-        # Cannot use self.draftstate_enabled here as there are subclasses
-        # that rely on get_object to determine the model
-        if isinstance(self.live_object, DraftStateMixin):
-            return self.live_object.get_latest_revision_as_object()
-        return self.live_object
+        return super().get_object(queryset)
 
     def get_page_subtitle(self):
         return str(self.object)
@@ -547,23 +540,13 @@ class EditView(
         Called after the form is successfully validated - saves the object to the db.
         Override this to implement custom save logic.
         """
-        commit = not self.draftstate_enabled
-        instance = self.form.save(commit=commit)
-        self.new_revision = None
+        instance = self.form.save()
 
         self.has_content_changes = self.form.has_changed()
-
-        # Save revision if the model inherits from RevisionMixin
-        if self.revision_enabled:
-            self.new_revision = instance.save_revision(
-                user=self.request.user,
-                changed=self.has_content_changes,
-            )
 
         log(
             instance=instance,
             action="wagtail.edit",
-            revision=self.new_revision,
             content_changed=self.has_content_changes,
         )
 
@@ -580,27 +563,7 @@ class EditView(
             )
         return redirect(self.get_success_url())
 
-    def publish_action(self):
-        hook_response = self.run_hook("before_publish", self.request, self.object)
-        if hook_response is not None:
-            return hook_response
-
-        self.new_revision.publish(user=self.request.user)
-
-        hook_response = self.run_hook("after_publish", self.request, self.object)
-        if hook_response is not None:
-            return hook_response
-
-        return None
-
     def get_success_message(self):
-        if self.draftstate_enabled and self.action == "publish":
-            if self.object.go_live_at and self.object.go_live_at > timezone.now():
-                return _("'%(object)s' updated and scheduled for publishing.") % {
-                    "object": self.object
-                }
-            return _("'%(object)s' updated and published.") % {"object": self.object}
-
         if self.success_message is None:
             return None
         return self.success_message % {"object": self.object}
@@ -617,37 +580,10 @@ class EditView(
             return None
         return self.error_message
 
-    def get_live_last_updated_info(self):
-        # DraftStateMixin is applied but object is not live
-        if self.draftstate_enabled and not self.object.live:
-            return None
-
-        revision = None
-        # DraftStateMixin is applied and object is live
-        if self.draftstate_enabled and self.object.live_revision:
-            revision = self.object.live_revision
-        # RevisionMixin is applied, so object is assumed to be live
-        elif self.revision_enabled and self.object.latest_revision:
-            revision = self.object.latest_revision
-
-        # No mixin is applied or no revision exists, fall back to latest log entry
-        if not revision:
-            return log_registry.get_logs_for_instance(self.object).first()
-
-        return {
-            "timestamp": revision.created_at,
-            "user_display_name": user_display_name(revision.user),
-        }
-
     def form_valid(self, form):
         self.form = form
         with transaction.atomic():
             self.object = self.save_instance()
-
-        if self.action == "publish" and self.draftstate_enabled:
-            response = self.publish_action()
-            if response is not None:
-                return response
 
         response = self.save_action()
 
@@ -675,11 +611,6 @@ class EditView(
         if context["can_delete"]:
             context["delete_url"] = self.get_delete_url()
             context["delete_item_label"] = self.delete_item_label
-
-        context["revision_enabled"] = self.revision_enabled
-        context["draftstate_enabled"] = self.draftstate_enabled
-
-        context["live_last_updated_info"] = self.get_live_last_updated_info()
         return context
